refactor grid classic-modern step 2
