+++ src/app/pages/available-certifications-page/available-certifications-page.component.html
@@ -1,17 +1,19 @@
 <div class="available-certs-container">
   
   @if (layoutService.layout() === 'moderno') {
-    <app-certifications-card-grid-modern>
+    <app-certifications-card-grid-modern
       [isLoading]="isLoading"
       [errorLoading]="errorLoading"
-      [certifications]="certifications">
+      [certifications]="certifications"
+      (retryLoad)="loadCertifications()">
     </app-certifications-card-grid-modern>
     
   } @else {
-    <app-certifications-card-grid-classic>
+    <app-certifications-card-grid-classic
       [isLoading]="isLoading"
       [errorLoading]="errorLoading"
-      [certifications]="certifications">
+      [certifications]="certifications"
+      (retryLoad)="loadCertifications()">
     </app-certifications-card-grid-classic>
   }
