+++ src/app/pages/certification-take-modern/certification-take-modern.component.scss
