+++ src/app/pages/certification-take-page/certification-take-page.component.scss
@@ -1,0 +1,2 @@
+.error-message {
+}
